fix: refactor namespaces and get tests passing thanks to @peschkaj
