+++ Makefile
@@ -8,7 +8,7 @@ golint:
 	golangci-lint run --skip-dirs api,space
 
 test: build
-	go test -v .
+	go test -v ./tests
 
 compose:
 	cd ./.docker && \
