+++ go.mod
@@ -4,9 +4,6 @@ go 1.19
 
 require (
 	github.com/gofrs/uuid v4.3.0+incompatible
-	github.com/google/uuid v1.3.0
-	github.com/jxskiss/base62 v1.1.0
-	github.com/pkg/errors v0.9.1
 )
 
 require (
